implemented TDS
Implemented the TDS_driver.c and its functionality in sensors. Redefined TDS ADC pin because it was pulling the sensor low for some reason. The nutrient and tds are on the same ADC which makes reading more difficult so I added ADC_driver.c for an easy interface. Also cleaned up some of the code.